+++ tests/test_utils.py
@@ -372,15 +372,17 @@ def test_cutline():
         ],
     }
 
+    # Check when using `boundless cutline`
+    # https://github.com/cogeotiff/rio-tiler/issues/585
     triangle_bounds = featureBounds(triangle_over_image_edge)
-
     with Reader(COG_RGB) as src:
-
         cutline = utils.create_cutline(
             src.dataset, triangle_over_image_edge, geometry_crs="epsg:4326"
         )
         data, mask = src.part(triangle_bounds, vrt_options={"cutline": cutline})
-        assert np.sum(mask == 255) == 5984
+        assert sum(mask[:, 0]) == 0  # first line
+        assert sum(mask[0, :]) == 0  # first column
+        assert sum(mask[:, -1]) == 0  # last column
 
 
 def test_parse_expression():