+++ CHANGES.md
@@ -1,4 +1,8 @@
 
+# Next (TBD)
+
+* enable `boundless` cutline (author @yellowcap, https://github.com/cogeotiff/rio-tiler/pull/586)
+
 # 4.1.9 (2023-02-28)
 
 * Automatically expand 2D numpy array to 3D when creating ImageData